Update vote count M/R to use Erlang
JS Map/reduce jobs are a fail. This commit includes changes which allows for the map/reduce for vote counts to be conducted using deployed erlang modules. The commit includes a makefile change to allow for the modules to be compiled and loaded into Riak. This commit is specific to my environment, but so is everything else so why not!

Modules are stored in the `apps/csd_core/riak_modules` folder to keep them separate from the application's source.

Two different types of map reduce are removed and one added. The counting of values when the user is logged in is combined into the same query as when the user is not logged in.