Use search to load snippets with paging
Instead of M/R we are now using Riak Search to pull snippets out. This returns the values in chronological order instead of reverse chron.

We have the data we need to deal with paging, but the UI doesn't yet support paging as we aren't doing anything with the result.

Also fixed up a small issue with the webmachine logger while I was at it given that we're now using the latest version of WM.