Pagination and map/reduce count fix
Enabled pagination on the web UI so that a user can see the full list of snippets that they have submitted in chunks of 10. They are ordered by the date they were submitted using the presort feature of Riak Search.

Fixed up an issue in the map/reduce counter for vote counts which was causing things to crash. Stupid mistake really!